+++ examples/with_bundler/init.js
@@ -1,10 +1,10 @@
-let Matey = require("../..");
-let matey = new Matey();
+const Matey = require("../..");
+const matey = new Matey();
 
 document.body.innerHTML = '<div id="test"></div>';
 
-let config = {
-  rmlMapperUrl: "http://localhost:4000/execute" // make sure an endpoint with this URL is active!
+const config = {
+  rmlMapperUrl: "http://172.23.43.193:4000/execute" // make sure an endpoint with this URL is active!
 };
 
 //for debug